// src/app/modules/transportation/models/vehicle-service-view.ts
/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AttachmentViewModel } from 'app/modules/academics/student-admissions/model/attachment-view-model';

/**
 * VehicleServiceView
 */
export interface VehicleServiceView { 
    /**
     * Gets Or Sets
     */
    vehicleId?: string;
    /**
     * Gets Or Sets
     */
    serviceDate?: Date;
    /**
     * Gets Or Sets
     */
    odometerReadingAtService?: number;
    /**
     * Gets Or Sets
     */
    nextServiceDate?: Date;
    /**
     * Gets Or Sets
     */
    odomenterReadingNextService?: number;
    /**
     * Gets Or Sets
     */
    additionalDetails?: string;
    /**
     * Gets Or Sets
     */
    attachment?: AttachmentViewModel;
}